harden and modularize nkey authentication
