+++ internal/Policy/Auth/NKey.hs
@@ -4,6 +4,7 @@ module Auth.NKey
   , signNonceWithSeed
   ) where
 
+import qualified Auth.NKey.Codec as Codec
 import           Auth.Types
 
 auth :: NKeyData -> Auth
@@ -12,3 +13,8 @@ auth seed = emptyAuth { authChallenge = Just (AuthNKeySeed seed) }
 authHandler :: NKeyPublicKey -> SignatureHandler -> Auth
 authHandler publicKey handler =
   emptyAuth { authChallenge = Just (AuthNKeyHandler publicKey handler) }
+
+signNonceWithSeed :: NKeyData -> Nonce -> Either String (NKeyPublicKey, Signature)
+signNonceWithSeed seed nonce = do
+  (publicKey, signature) <- Codec.signNonceWithSeedRaw seed nonce
+  pure (publicKey, Codec.encodeSignature signature)
