+++ natskell.cabal
@@ -112,7 +112,7 @@ library natskell-internal
     time >= 1.9 && < 1.16,
     network  >= 3.1 && < 3.3,
     mtl >= 2.2 && < 2.4,
-    cryptonite >= 0.29 && < 0.31,
+    crypton >= 0.34 && < 1.1,
     memory >= 0.14 && < 0.19,
     random  >= 1.1 && < 1.3,
     network-simple  >= 0.4 && < 0.5,
@@ -124,6 +124,8 @@ library natskell-internal
   exposed-modules:
     Client.API
     Auth.Types
+    Auth.Config
+    Auth.Resolver
     Auth.None
     Auth.Token
     Auth.UserPass
@@ -179,6 +181,9 @@ library natskell-internal
     Network.Connection.Tls
     Queue.API
     Queue.TransactionalQueue
+  other-modules:
+    Auth.Credentials
+    Auth.NKey.Codec
   hs-source-dirs:
     internal
     internal/Types
